backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
pydantic-settings==2.8.1
sqlalchemy==2.0.39
alembic==1.15.2
psycopg[binary]==3.2.6
PyYAML==6.0.2
redis==5.2.1
chromadb==1.5.3
nats-py==2.11.0
cryptography==46.0.7
email-validator==2.2.0
pytest==8.3.5
httpx==0.28.1
dingtalk-stream==0.24.3
websockets==13.1
